Updated examples to use the new waitForTransactionFinalization
