+++ examples/common/deployModule.ts
@@ -4,13 +4,10 @@ import {
     AccountTransactionHeader,
     AccountTransactionSignature,
     AccountTransactionType,
-    buildBasicAccountSigner,
     DeployModulePayload,
     signTransaction,
     TransactionExpiry,
     createConcordiumClient,
-    unwrap,
-    HexString,
     parseWallet,
     buildAccountSigner,
 } from '@concordium/node-sdk';
@@ -107,13 +104,7 @@ const client = createConcordiumClient(
     );
 
     console.log('Transaction submitted, waiting for finalization...');
-    await client.waitForTransactionFinalization(transactionHash);
 
-    console.log('Transaction finalized, getting outcome...\n');
-
-    const transactionStatus = await client.getBlockItemStatus(transactionHash);
-
-    if (transactionStatus.status === 'finalized') {
-        console.dir(transactionStatus.outcome, { depth: null, colors: true });
-    }
+    const status = await client.waitForTransactionFinalization(transactionHash);
+    console.dir(status, { depth: null, colors: true });
 })();
